refactor(process): migrate rtcontainercheck component to v2
