+++ comp/process/bundle.go
@@ -15,16 +15,16 @@ import (
 	connectionsforwarderfx "github.com/DataDog/datadog-agent/comp/forwarder/connectionsforwarder/fx"
 	"github.com/DataDog/datadog-agent/comp/process/agent/agentimpl"
 	"github.com/DataDog/datadog-agent/comp/process/apiserver"
-	connectionscheckfx "github.com/DataDog/datadog-agent/comp/process/connectionscheck/fx"
+	"github.com/DataDog/datadog-agent/comp/process/connectionscheck/connectionscheckimpl"
 	"github.com/DataDog/datadog-agent/comp/process/containercheck/containercheckimpl"
 	expvarsfx "github.com/DataDog/datadog-agent/comp/process/expvars/fx"
 	forwardersfx "github.com/DataDog/datadog-agent/comp/process/forwarders/fx"
 	gpusubscriber "github.com/DataDog/datadog-agent/comp/process/gpusubscriber/fx"
 	hostinfofx "github.com/DataDog/datadog-agent/comp/process/hostinfo/fx"
 	processcheckfx "github.com/DataDog/datadog-agent/comp/process/processcheck/fx"
-	processdiscoverycheckfx "github.com/DataDog/datadog-agent/comp/process/processdiscoverycheck/fx"
+	"github.com/DataDog/datadog-agent/comp/process/processdiscoverycheck/processdiscoverycheckimpl"
 	profilerimpl "github.com/DataDog/datadog-agent/comp/process/profiler/fx"
-	"github.com/DataDog/datadog-agent/comp/process/rtcontainercheck/rtcontainercheckimpl"
+	rtcontainercheckfx "github.com/DataDog/datadog-agent/comp/process/rtcontainercheck/fx"
 	"github.com/DataDog/datadog-agent/comp/process/runner/runnerimpl"
 	submitterfx "github.com/DataDog/datadog-agent/comp/process/submitter/fx"
 	logscompression "github.com/DataDog/datadog-agent/comp/serializer/logscompression/fx"
@@ -43,11 +43,11 @@ func Bundle() fxutil.BundleOptions {
 		profilerimpl.Module(),
 
 		// Checks
-		connectionscheckfx.Module(),
+		connectionscheckimpl.Module(),
 		containercheckimpl.Module(),
 		processcheckfx.Module(),
-		rtcontainercheckimpl.Module(),
-		processdiscoverycheckfx.Module(),
+		rtcontainercheckfx.Module(),
+		processdiscoverycheckimpl.Module(),
 
 		agentimpl.Module(),
