+++ comp/process/rtcontainercheck/def/component.go
